+++ SteamStorageAPI/Services/Domain/StatisticsService/StatisticsService.cs
@@ -33,19 +33,19 @@ public class StatisticsService : IStatisticsService
         User user,
         CancellationToken cancellationToken = default)
     {
-        double rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
+        decimal rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
 
         IQueryable<Active> actives = GetActivesQuery(user);
         IQueryable<Archive> archives = GetArchivesQuery(user);
 
-        double investedSum =
-            (double)(actives.Sum(y => y.BuyPrice * y.Count) + archives.Sum(y => y.BuyPrice * y.Count));
+        decimal investedSum =
+            actives.Sum(y => y.BuyPrice * y.Count) + archives.Sum(y => y.BuyPrice * y.Count);
 
-        double currentSum =
-            (double)actives.Sum(x => x.Skin.CurrentPrice * x.Count) * rate
-            + (double)archives.Sum(y => y.SoldPrice * y.Count);
+        decimal currentSum =
+            actives.Sum(x => x.Skin.CurrentPrice * x.Count) * rate
+            + archives.Sum(y => y.SoldPrice * y.Count);
 
-        double percentage = investedSum == 0 ? 1 : (currentSum - investedSum) / investedSum;
+        decimal percentage = investedSum == 0 ? 1 : (currentSum - investedSum) / investedSum;
 
         return new InvestmentSumResponse(currentSum, percentage);
     }
@@ -54,18 +54,18 @@ public class StatisticsService : IStatisticsService
         User user,
         CancellationToken cancellationToken = default)
     {
-        double rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
+        decimal rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
 
         IQueryable<Active> actives = GetActivesQuery(user);
         IQueryable<Archive> archives = GetArchivesQuery(user);
 
-        double financialGoal = (double)(user.GoalSum ?? 0);
+        decimal financialGoal = user.GoalSum ?? 0;
 
-        double currentSum =
-            (double)actives.Sum(x => x.Skin.CurrentPrice * x.Count) * rate
-            + (double)archives.Sum(y => y.SoldPrice * y.Count);
+        decimal currentSum =
+            actives.Sum(x => x.Skin.CurrentPrice * x.Count) * rate
+            + archives.Sum(y => y.SoldPrice * y.Count);
 
-        double percentage = financialGoal == 0 ? 1 : currentSum / financialGoal;
+        decimal percentage = financialGoal == 0 ? 1 : currentSum / financialGoal;
 
         return new FinancialGoalResponse(financialGoal, percentage);
     }
@@ -74,14 +74,14 @@ public class StatisticsService : IStatisticsService
         User user,
         CancellationToken cancellationToken = default)
     {
-        double rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
+        decimal rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
 
         IQueryable<Active> actives = GetActivesQuery(user);
 
         int count = actives.Sum(x => x.Count);
-        double investedSum = (double)actives.Sum(y => y.BuyPrice * y.Count);
-        double currentSum = (double)actives.Sum(x => x.Skin.CurrentPrice * x.Count) * rate;
-        double percentage = investedSum == 0 ? 1 : (currentSum - investedSum) / investedSum;
+        decimal investedSum = actives.Sum(y => y.BuyPrice * y.Count);
+        decimal currentSum = actives.Sum(x => x.Skin.CurrentPrice * x.Count) * rate;
+        decimal percentage = investedSum == 0 ? 1 : (currentSum - investedSum) / investedSum;
 
         return new ActiveStatisticResponse(count, currentSum, percentage);
     }
@@ -93,9 +93,9 @@ public class StatisticsService : IStatisticsService
         IQueryable<Archive> archives = GetArchivesQuery(user);
 
         int count = archives.Sum(x => x.Count);
-        double investedSum = (double)archives.Sum(y => y.BuyPrice * y.Count);
-        double soldSum = (double)archives.Sum(y => y.SoldPrice * y.Count);
-        double percentage = investedSum == 0 ? 1 : (soldSum - investedSum) / investedSum;
+        decimal investedSum = archives.Sum(y => y.BuyPrice * y.Count);
+        decimal soldSum = archives.Sum(y => y.SoldPrice * y.Count);
+        decimal percentage = investedSum == 0 ? 1 : (soldSum - investedSum) / investedSum;
 
         return new ArchiveStatisticResponse(count, soldSum, percentage);
     }
@@ -104,7 +104,7 @@ public class StatisticsService : IStatisticsService
         User user,
         CancellationToken cancellationToken = default)
     {
-        double rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
+        decimal rate = await _currencyService.GetCurrencyExchangeRateAsync(user, cancellationToken);
 
         List<Inventory> inventories = await _context.Entry(user)
             .Collection(u => u.Inventories)
@@ -114,7 +114,7 @@ public class StatisticsService : IStatisticsService
             .ToListAsync(cancellationToken);
 
         int count = inventories.Sum(x => x.Count);
-        double sum = (double)inventories.Sum(x => x.Skin.CurrentPrice * x.Count) * rate;
+        decimal sum = inventories.Sum(x => x.Skin.CurrentPrice * x.Count) * rate;
 
         List<Game> games = inventories
             .Select(x => x.Skin.Game)
@@ -124,7 +124,7 @@ public class StatisticsService : IStatisticsService
         List<InventoryGameStatisticResponse> gamesResponse = games.Select(item =>
             new InventoryGameStatisticResponse(
                 item.Title,
-                count == 0 ? 0 : (double)inventories.Where(x => x.Skin.GameId == item.Id).Sum(x => x.Count) / count,
+                count == 0 ? 0 : inventories.Where(x => x.Skin.GameId == item.Id).Sum(x => x.Count) / (decimal)count,
                 inventories.Where(x => x.Skin.GameId == item.Id).Sum(x => x.Count))).ToList();
 
         return new InventoryStatisticResponse(count, sum, gamesResponse);
